Updated Readme, added loader in API calls, added responsiveness

+++ client/src/components/ListItem.js
@@ -5,8 +5,12 @@ import ProgressBar from './ProgressBar'
 
 const ListItem = ({ task, getData }) => {
 
+  const [showItem, setShowItem] = useState(true);
+
   const deleteItem = async () => {
     try {
+
+      setShowItem(false);
 
       const response = await fetch(`${process.env.REACT_APP_SERVERURL}/todos/${task.id}`,
         {
@@ -21,21 +25,25 @@ const ListItem = ({ task, getData }) => {
     }
   }
   const [showModal, setShowModal] = useState(false);
-  return (
-    <div className="list-item">
-      <div className="info-container">
-        <TickIcon progress={task.progress} />
-        <p className="task-title">{task.title}</p>
-        <ProgressBar progress={task.progress} />
-      </div>
+  if (showItem) {
+    return (
+      <div className="list-item">
+        <div className="info-container">
+          <TickIcon progress={task.progress} />
+          <p className="task-title">{task.title}</p>
+          <ProgressBar progress={task.progress} />
+        </div>
 
-      <div className="button-container">
-        <button className="edit" onClick={() => setShowModal(true)}>EDIT</button>
-        <button className="delete" onClick={deleteItem}>DELETE</button>
+        <div className="button-container">
+          <button className="edit" onClick={() => setShowModal(true)}>EDIT</button>
+          <button className="delete" onClick={deleteItem}>DELETE</button>
+        </div>
+        {showModal && <Modal mode={'edit'} setShowModal={setShowModal} getData={getData} task={task} />}
       </div>
-      {showModal && <Modal mode={'edit'} setShowModal={setShowModal} getData={getData} task={task} />}
-    </div>
-  )
+    )
+  }
+  else return (<></>);
+
 }
 
 export default ListItem
